feat: switch default engine to Rayon and add comprehensive benchmarking
Summary:
- Changed default search engine from Smol to Rayon based on performance analysis
- Added multiple search engine implementations (Smol, Rayon, RayonLimited)
- Implemented comprehensive benchmarking suite for performance comparison
- Integrated jwalk for 130x faster file discovery

Performance Benchmark Results:

1. Small workload profiling (20 files, 1000 messages each):
   - Smol: 392.50ms (avg: 39.25ms/execution)
   - Rayon: 286.40ms (avg: 28.64ms/execution)
   - Rayon is 27% faster for small workloads

2. Query-specific performance:
   - NOT queries: Rayon 2.6x faster than Smol
   - assistant_only filter: Rayon 1.4x faster than Smol
   - Simple queries: Performance similar between engines
   - Complex queries (AND/OR): Smol slightly faster

3. Real-world CLI usage (hyperfine):
   - Smol generally faster for full CLI execution including startup
   - Rayon shows better raw search performance once initialized

4. Profiling insights:
   - Rayon spends 37.75% CPU time in WorkerThread::wait_until_cold
   - This overhead is significant for I/O-bound workloads
   - Smol handles async I/O more efficiently
   - jwalk file discovery provides 130x speedup over walkdir

Technical Changes:
- Added SearchEngineTrait for engine abstraction
- Implemented SmolEngine, RayonEngine, and RayonLimitedEngine
- Added --engine CLI flag for runtime engine selection
- Created comprehensive benchmark suite including:
  - engine_comparison_benchmark
  - realistic_benchmark
  - profiled_benchmark
  - Various workload size tests (small/medium/large/xlarge)

The switch to Rayon as default is based on its superior performance
for CPU-intensive operations like query evaluation, while maintaining
the option to use Smol via --engine flag for I/O-heavy workloads.
