+++ benches/large_file_benchmark.rs
@@ -1,4 +1,4 @@
-use ccms::{SearchEngine, SearchOptions, parse_query};
+use ccms::{SearchEngineTrait, SmolEngine, SearchOptions, parse_query};
 use codspeed_criterion_compat::{Criterion, black_box, criterion_group, criterion_main};
 use std::fs::File;
 use std::io::Write;
@@ -27,7 +27,7 @@ fn benchmark_large_file_search(c: &mut Criterion) {
 
     c.bench_function("search_10k_lines", |b| {
         b.iter(|| {
-            let engine = SearchEngine::new(options.clone());
+            let engine = SmolEngine::new(options.clone());
             let (results, _, _) = engine.search(&test_file, black_box(query.clone())).unwrap();
             results
         });
@@ -41,7 +41,7 @@ fn benchmark_very_large_file_search(c: &mut Criterion) {
 
     c.bench_function("search_50k_lines", |b| {
         b.iter(|| {
-            let engine = SearchEngine::new(options.clone());
+            let engine = SmolEngine::new(options.clone());
             let (results, _, _) = engine.search(&test_file, black_box(query.clone())).unwrap();
             results
         });
