+++ claude_rayon_comprehensive.txt
@@ -1,0 +1,61 @@
+=== Enhanced Profiling Report: main ===
+Total runtime: 0.320s
+
+CPU Profiling Report (pprof)
+============================
+
+Top Functions by CPU Time:
+--------------------------
+  1.  66.67% (   810 samples, ~ 810.0ms) std::io::read_until
+  2.  43.13% (   524 samples, ~   0.0ms) rayon_registry::WorkerThread::wait_until_cold
+  3.  41.98% (   510 samples, ~  27.0ms) <rayon_job::HeapJob<BODY> as rayon_job::Job>::execute
+  4.  40.58% (   493 samples, ~   0.0ms) std::sys::backtrace::__rust_begin_short_backtrace
+  5.  40.49% (   492 samples, ~   0.0ms) std::sys::pal::unix::thread::Thread::new::thread_start
+  6.  40.49% (   492 samples, ~   0.0ms) ops::function::FnOnce::call_once{{vtable.shim}}
+  7.   9.71% (   118 samples, ~  47.0ms) <&mut sonic_rs::serde::de::Deserializer<R> as serde::de::Deserializer>::deserialize_any
+  8.   5.93% (    72 samples, ~  72.0ms) alloc::str::<impl str>::to_lowercase
+  9.   4.12% (    50 samples, ~  43.0ms) sonic_rs::parser::Parser<R>::parse_string_escaped
+ 10.   3.79% (    46 samples, ~  11.0ms) ccms::query::condition::QueryCondition::evaluate
+ 11.   3.29% (    40 samples, ~   0.0ms) <rayon_job::StackJob<L,F,R> as rayon_job::Job>::execute
+ 12.   2.88% (    35 samples, ~   6.0ms) ccms::schemas::session_message::_::<impl serde::de::Deserialize for ccms::schemas::session_messag...
+ 13.   2.63% (    32 samples, ~   0.0ms) rayon_scope::scope::{{closure}}
+ 14.   2.55% (    31 samples, ~   1.0ms) sonic_rs::serde::de::from_trait
+ 15.   2.30% (    28 samples, ~  25.0ms) __mi_malloc_generic
+ 16.   2.30% (    28 samples, ~  28.0ms) _mi_find_page
+ 17.   1.98% (    24 samples, ~  24.0ms) _posix_spawnattr_set_qos_clamp_np
+ 18.   1.73% (    21 samples, ~   0.0ms) rayon::iter::plumbing::bridge_unindexed_producer_consumer
+ 19.   1.48% (    18 samples, ~  15.0ms) __mi_heap_realloc_zero
+ 20.   1.40% (    17 samples, ~  15.0ms) <serde::__private::de::content::ContentDeserializer<E> as serde::de::Deserializer>::deserialize_any
+ 21.   1.07% (    13 samples, ~  13.0ms) _mi_page_fresh_alloc
+ 22.   1.07% (    13 samples, ~  13.0ms) _os_cpu_in_cksum_mbuf
+ 23.   0.91% (    11 samples, ~   0.0ms) rayon_join::join_context::{{closure}}
+ 24.   0.91% (    11 samples, ~   1.0ms) std::sys::fs::read_dir
+ 25.   0.82% (    10 samples, ~   0.0ms) <jwalk::WalkDirGeneric<C> as iter::traits::collect::IntoIterator>::into_iter::{{closure}}
+ 26.   0.82% (    10 samples, ~   0.0ms) <&rayon::iter::par_bridge::IterParallelProducer<Iter> as rayon::iter::plumbing::UnindexedProducer...
+ 27.   0.58% (     7 samples, ~   7.0ms) sonic_rs::reader::Read::new
+ 28.   0.58% (     7 samples, ~   7.0ms) _mi_malloc_aligned
+ 29.   0.58% (     7 samples, ~   1.0ms) alloc::raw_vec::RawVecInner<A>::reserve::do_reserve_and_handle
+ 30.   0.49% (     6 samples, ~   1.0ms) <serde::__private::de::content::ContentVisitor as serde::de::Visitor>::visit_seq
+ 31.   0.49% (     6 samples, ~   0.0ms) alloc::raw_vec::finish_grow
+ 32.   0.49% (     6 samples, ~   6.0ms) alloc::str::join_generic_copy
+ 33.   0.41% (     5 samples, ~   5.0ms) serde::de::impls::<impl serde::de::Deserialize for alloc::string::String>::deserialize
+ 34.   0.41% (     5 samples, ~   2.0ms) <marker::PhantomData<T> as serde::de::DeserializeSeed>::deserialize
+ 35.   0.41% (     5 samples, ~   0.0ms) std::fs::OpenOptions::_open
+ 36.   0.33% (     4 samples, ~   4.0ms) ccms::schemas::session_message::SessionMessage::get_content_text
+ 37.   0.33% (     4 samples, ~   2.0ms) <serde::__private::de::content::ContentRefDeserializer<E> as serde::de::Deserializer>::deserializ...
+ 38.   0.25% (     3 samples, ~   2.0ms) ccms::schemas::session_message::SessionMessage::get_searchable_text
+ 39.   0.25% (     3 samples, ~   0.0ms) _mi_heap_malloc_zero_aligned_at_generic
+ 40.   0.16% (     2 samples, ~   0.0ms) <serde::de::impls::OptionVisitor<T> as serde::de::Visitor>::visit_some
+ 41.   0.16% (     2 samples, ~   2.0ms) std::sys::fs::metadata
+ 42.   0.16% (     2 samples, ~   2.0ms) <&mut A as serde::de::MapAccess>::next_value
+ 43.   0.16% (     2 samples, ~   2.0ms) alloc::collections::btree::map::BTreeMap<K,V,A>::insert
+ 44.   0.16% (     2 samples, ~   0.0ms) jwalk::read_dir_iter::ReadDirIter<C>::try_new::{{closure}}
+ 45.   0.16% (     2 samples, ~   2.0ms) <serde::__private::de::FlatMapDeserializer<E> as serde::de::Deserializer>::deserialize_struct
+ 46.   0.16% (     2 samples, ~   2.0ms) <std::fs::ReadDir as iter::traits::iterator::Iterator>::next
+ 47.   0.16% (     2 samples, ~   1.0ms) <serde::__private::de::content::Content as clone::Clone>::clone
+ 48.   0.16% (     2 samples, ~   2.0ms) _mi_free
+ 49.   0.16% (     2 samples, ~   2.0ms) serde::de::value::MapDeserializer<I,E>::end
+
+Total samples: 1215
+Sampling frequency: 1000 Hz
+Estimated CPU time: 1.215s
