+++ claude_smol_comprehensive.txt
@@ -1,0 +1,62 @@
+=== Enhanced Profiling Report: main ===
+Total runtime: 0.406s
+
+CPU Profiling Report (pprof)
+============================
+
+Top Functions by CPU Time:
+--------------------------
+  1.  62.25% (   798 samples, ~ 798.0ms) std::io::read_until
+  2.  41.97% (   538 samples, ~  44.0ms) blocking::unblock::{{closure}}
+  3.  39.08% (   501 samples, ~   0.0ms) std::sys::backtrace::__rust_begin_short_backtrace
+  4.  38.85% (   498 samples, ~   0.0ms) std::sys::pal::unix::thread::Thread::new::thread_start
+  5.  38.85% (   498 samples, ~   0.0ms) ops::function::FnOnce::call_once{{vtable.shim}}
+  6.  38.53% (   494 samples, ~   0.0ms) async_task::raw::RawTask<F,T,S,M>::run
+  7.   9.44% (   121 samples, ~  46.0ms) <&mut sonic_rs::serde::de::Deserializer<R> as serde::de::Deserializer>::deserialize_any
+  8.   5.62% (    72 samples, ~  72.0ms) alloc::str::<impl str>::to_lowercase
+  9.   4.84% (    62 samples, ~  56.0ms) sonic_rs::parser::Parser<R>::parse_string_escaped
+ 10.   4.52% (    58 samples, ~  18.0ms) ccms::query::condition::QueryCondition::evaluate
+ 11.   3.12% (    40 samples, ~  38.0ms) _mi_find_page
+ 12.   2.73% (    35 samples, ~  23.0ms) __mi_malloc_generic
+ 13.   2.34% (    30 samples, ~   5.0ms) ccms::schemas::session_message::_::<impl serde::de::Deserialize for ccms::schemas::session_messag...
+ 14.   2.03% (    26 samples, ~   1.0ms) sonic_rs::serde::de::from_trait
+ 15.   1.64% (    21 samples, ~  17.0ms) __mi_heap_realloc_zero
+ 16.   1.64% (    21 samples, ~  21.0ms) _mi_page_fresh_alloc
+ 17.   1.56% (    20 samples, ~  20.0ms) _os_cpu_in_cksum_mbuf
+ 18.   1.33% (    17 samples, ~  15.0ms) <serde::__private::de::content::ContentDeserializer<E> as serde::de::Deserializer>::deserialize_any
+ 19.   1.17% (    15 samples, ~   7.0ms) alloc::str::join_generic_copy
+ 20.   0.94% (    12 samples, ~   0.0ms) _mi_heap_malloc_zero_aligned_at_generic
+ 21.   0.86% (    11 samples, ~   3.0ms) ccms::schemas::session_message::SessionMessage::get_content_text
+ 22.   0.70% (     9 samples, ~   3.0ms) alloc::raw_vec::RawVecInner<A>::reserve::do_reserve_and_handle
+ 23.   0.62% (     8 samples, ~   8.0ms) sonic_rs::reader::Read::new
+ 24.   0.62% (     8 samples, ~   8.0ms) _mi_malloc_aligned
+ 25.   0.55% (     7 samples, ~   1.0ms) <serde::__private::de::content::ContentVisitor as serde::de::Visitor>::visit_seq
+ 26.   0.55% (     7 samples, ~   0.0ms) std::fs::OpenOptions::_open
+ 27.   0.55% (     7 samples, ~   7.0ms) <std::fs::ReadDir as iter::traits::iterator::Iterator>::next
+ 28.   0.55% (     7 samples, ~   5.0ms) <marker::PhantomData<T> as serde::de::DeserializeSeed>::deserialize
+ 29.   0.55% (     7 samples, ~   7.0ms) serde::de::impls::<impl serde::de::Deserialize for alloc::string::String>::deserialize
+ 30.   0.55% (     7 samples, ~   7.0ms) <serde::__private::de::FlatMapDeserializer<E> as serde::de::Deserializer>::deserialize_struct
+ 31.   0.55% (     7 samples, ~   7.0ms) _posix_spawnattr_set_qos_clamp_np
+ 32.   0.47% (     6 samples, ~   0.0ms) alloc::raw_vec::finish_grow
+ 33.   0.47% (     6 samples, ~   4.0ms) <serde::__private::de::content::ContentRefDeserializer<E> as serde::de::Deserializer>::deserializ...
+ 34.   0.39% (     5 samples, ~   2.0ms) ccms::main
+ 35.   0.31% (     4 samples, ~   4.0ms) ptr::drop_in_place<serde_json::value::Value>
+ 36.   0.31% (     4 samples, ~   1.0ms) <search::smol_engine::SmolEngine as search::engine::SearchEngineTrait>::search_with_role_filter_a...
+ 37.   0.23% (     3 samples, ~   1.0ms) <serde::de::impls::OptionVisitor<T> as serde::de::Visitor>::visit_some
+ 38.   0.23% (     3 samples, ~   0.0ms) rayon::iter::plumbing::bridge_unindexed_producer_consumer
+ 39.   0.23% (     3 samples, ~   3.0ms) serde::__private::de::flat_map_take_entry
+ 40.   0.23% (     3 samples, ~   0.0ms) <rayon_job::StackJob<L,F,R> as rayon_job::Job>::execute
+ 41.   0.23% (     3 samples, ~   3.0ms) _mi_free
+ 42.   0.23% (     3 samples, ~   0.0ms) <&rayon::iter::par_bridge::IterParallelProducer<Iter> as rayon::iter::plumbing::UnindexedProducer...
+ 43.   0.23% (     3 samples, ~   0.0ms) <jwalk::WalkDirGeneric<C> as iter::traits::collect::IntoIterator>::into_iter::{{closure}}
+ 44.   0.23% (     3 samples, ~   0.0ms) rayon_registry::WorkerThread::wait_until_cold
+ 45.   0.23% (     3 samples, ~   0.0ms) search::file_discovery::FileDiscovery::discover_files
+ 46.   0.23% (     3 samples, ~   0.0ms) search::file_discovery::discover_claude_files
+ 47.   0.23% (     3 samples, ~   3.0ms) _mi_os_decommit_ex
+ 48.   0.23% (     3 samples, ~   0.0ms) _main
+ 49.   0.16% (     2 samples, ~   0.0ms) <iter::adapters::map::Map<I,F> as iter::traits::iterator::Iterator>::next
+ 50.   0.16% (     2 samples, ~   0.0ms) std::sys::fs::read_dir
+
+Total samples: 1282
+Sampling frequency: 1000 Hz
+Estimated CPU time: 1.282s
