+++ not_smol_comprehensive.txt
@@ -1,0 +1,62 @@
+=== Enhanced Profiling Report: main ===
+Total runtime: 0.401s
+
+CPU Profiling Report (pprof)
+============================
+
+Top Functions by CPU Time:
+--------------------------
+  1.  41.90% (   603 samples, ~ 603.0ms) std::io::read_until
+  2.  38.22% (   550 samples, ~ 133.0ms) blocking::unblock::{{closure}}
+  3.  36.07% (   519 samples, ~   0.0ms) std::sys::backtrace::__rust_begin_short_backtrace
+  4.  34.12% (   491 samples, ~   0.0ms) ops::function::FnOnce::call_once{{vtable.shim}}
+  5.  34.12% (   491 samples, ~   0.0ms) std::sys::pal::unix::thread::Thread::new::thread_start
+  6.  33.01% (   475 samples, ~   0.0ms) async_task::raw::RawTask<F,T,S,M>::run
+  7.  13.76% (   198 samples, ~ 191.0ms) _mi_find_page
+  8.   7.09% (   102 samples, ~  31.0ms) ccms::query::condition::QueryCondition::evaluate
+  9.   6.67% (    96 samples, ~  36.0ms) <&mut sonic_rs::serde::de::Deserializer<R> as serde::de::Deserializer>::deserialize_any
+ 10.   4.86% (    70 samples, ~  70.0ms) _mi_page_fresh_alloc
+ 11.   4.73% (    68 samples, ~   7.0ms) __mi_malloc_generic
+ 12.   4.10% (    59 samples, ~  50.0ms) sonic_rs::parser::Parser<R>::parse_string_escaped
+ 13.   3.96% (    57 samples, ~   1.0ms) _mi_heap_malloc_zero_aligned_at_generic
+ 14.   3.96% (    57 samples, ~  57.0ms) alloc::str::<impl str>::to_lowercase
+ 15.   2.92% (    42 samples, ~  11.0ms) alloc::str::join_generic_copy
+ 16.   2.64% (    38 samples, ~  38.0ms) _posix_spawnattr_set_qos_clamp_np
+ 17.   2.57% (    37 samples, ~  16.0ms) <search::smol_engine::SmolEngine as search::engine::SearchEngineTrait>::search_with_role_filter_a...
+ 18.   2.50% (    36 samples, ~   0.0ms) rayon::iter::plumbing::bridge_unindexed_producer_consumer
+ 19.   2.43% (    35 samples, ~   8.0ms) ccms::schemas::session_message::_::<impl serde::de::Deserialize for ccms::schemas::session_messag...
+ 20.   2.29% (    33 samples, ~   7.0ms) ccms::schemas::session_message::SessionMessage::get_content_text
+ 21.   1.95% (    28 samples, ~   0.0ms) _main
+ 22.   1.95% (    28 samples, ~   0.0ms) ccms::main
+ 23.   1.88% (    27 samples, ~   0.0ms) sonic_rs::serde::de::from_trait
+ 24.   1.81% (    26 samples, ~  21.0ms) __mi_heap_realloc_zero
+ 25.   1.67% (    24 samples, ~  16.0ms) slice::sort::stable::quicksort::quicksort
+ 26.   1.46% (    21 samples, ~   0.0ms) rayon_join::join_context::{{closure}}
+ 27.   1.39% (    20 samples, ~   8.0ms) <serde::__private::de::content::ContentDeserializer<E> as serde::de::Deserializer>::deserialize_any
+ 28.   1.25% (    18 samples, ~  18.0ms) ptr::drop_in_place<ccms::query::condition::SearchResult>
+ 29.   1.18% (    17 samples, ~  14.0ms) slice::sort::stable::drift::sort
+ 30.   1.11% (    16 samples, ~   4.0ms) ccms::schemas::session_message::SessionMessage::get_searchable_text
+ 31.   1.11% (    16 samples, ~   1.0ms) std::sys::fs::read_dir
+ 32.   1.04% (    15 samples, ~   0.0ms) <jwalk::WalkDirGeneric<C> as iter::traits::collect::IntoIterator>::into_iter::{{closure}}
+ 33.   1.04% (    15 samples, ~   0.0ms) rayon_registry::WorkerThread::wait_until_cold
+ 34.   1.04% (    15 samples, ~   0.0ms) <&rayon::iter::par_bridge::IterParallelProducer<Iter> as rayon::iter::plumbing::UnindexedProducer...
+ 35.   0.83% (    12 samples, ~   0.0ms) <rayon_job::StackJob<L,F,R> as rayon_job::Job>::execute
+ 36.   0.83% (    12 samples, ~   4.0ms) <serde::__private::de::content::ContentVisitor as serde::de::Visitor>::visit_seq
+ 37.   0.76% (    11 samples, ~   3.0ms) <marker::PhantomData<T> as serde::de::DeserializeSeed>::deserialize
+ 38.   0.76% (    11 samples, ~   4.0ms) <serde::__private::de::content::ContentRefDeserializer<E> as serde::de::Deserializer>::deserializ...
+ 39.   0.69% (    10 samples, ~   0.0ms) slice::sort::stable::driftsort_main
+ 40.   0.69% (    10 samples, ~   9.0ms) serde::de::impls::<impl serde::de::Deserialize for alloc::string::String>::deserialize
+ 41.   0.63% (     9 samples, ~   9.0ms) _mi_malloc_aligned
+ 42.   0.63% (     9 samples, ~   9.0ms) _os_cpu_in_cksum_mbuf
+ 43.   0.56% (     8 samples, ~   0.0ms) alloc::raw_vec::finish_grow
+ 44.   0.49% (     7 samples, ~   6.0ms) <serde::__private::de::FlatMapDeserializer<E> as serde::de::Deserializer>::deserialize_struct
+ 45.   0.49% (     7 samples, ~   0.0ms) <serde::de::impls::OptionVisitor<T> as serde::de::Visitor>::visit_some
+ 46.   0.35% (     5 samples, ~   1.0ms) alloc::raw_vec::RawVecInner<A>::reserve::do_reserve_and_handle
+ 47.   0.35% (     5 samples, ~   1.0ms) alloc::raw_vec::RawVec<T,A>::grow_one
+ 48.   0.35% (     5 samples, ~   2.0ms) <async_executor::AsyncCallOnDrop<Fut,Cleanup> as future::future::Future>::poll
+ 49.   0.35% (     5 samples, ~   5.0ms) sonic_rs::reader::Read::new
+ 50.   0.28% (     4 samples, ~   4.0ms) <std::fs::ReadDir as iter::traits::iterator::Iterator>::next
+
+Total samples: 1439
+Sampling frequency: 1000 Hz
+Estimated CPU time: 1.439s
