+++ src/main.rs
@@ -4,7 +4,7 @@ static GLOBAL: mimalloc::MiMalloc = mimalloc::MiMalloc;
 
 use anyhow::Result;
 use ccms::{
-    SearchEngine, SearchOptions, default_claude_pattern, format_search_result,
+    SearchEngineTrait, SmolEngine, RayonEngine, SearchOptions, default_claude_pattern, format_search_result,
     interactive_ratatui::InteractiveSearch, parse_query, profiling,
 };
 #[cfg(feature = "profiling")]
@@ -98,6 +98,10 @@ struct Cli {
     /// Generate shell completion script
     #[arg(long = "completion", value_enum)]
     generator: Option<Shell>,
+
+    /// Search engine to use
+    #[arg(long, value_enum, default_value = "rayon")]
+    engine: EngineType,
 }
 
 #[derive(Clone, Copy, Debug, ValueEnum)]
@@ -105,6 +109,12 @@ enum OutputFormat {
     Text,
     Json,
     JsonL,
+}
+
+#[derive(Clone, Copy, Debug, ValueEnum)]
+enum EngineType {
+    Smol,
+    Rayon,
 }
 
 fn print_completions<G: Generator>(generator: G, cmd: &mut Command) {
@@ -223,10 +233,23 @@ fn main() -> Result<()> {
 
     // Execute search
     if cli.verbose {
-        eprintln!("Using optimized search engine");
+        eprintln!("Using {} engine", match cli.engine {
+            EngineType::Smol => "Smol",
+            EngineType::Rayon => "Rayon",
+        });
     }
-    let engine = SearchEngine::new(options);
-    let (results, duration, total_count) = engine.search(pattern_to_use, query)?;
+    
+    // Create appropriate engine based on CLI flag
+    let (results, duration, total_count) = match cli.engine {
+        EngineType::Smol => {
+            let engine = SmolEngine::new(options);
+            engine.search(pattern_to_use, query)?
+        }
+        EngineType::Rayon => {
+            let engine = RayonEngine::new(options);
+            engine.search(pattern_to_use, query)?
+        }
+    };
 
     // Output results
     let stdout = io::stdout();
