+++ src/search/mod.rs
@@ -1,5 +1,11 @@
 pub mod engine;
 pub mod file_discovery;
+pub mod smol_engine;
+pub mod rayon_engine;
+pub mod rayon_limited_engine;
 
-pub use engine::{SearchEngine, format_search_result};
+pub use engine::{SearchEngineTrait, format_search_result};
+pub use smol_engine::SmolEngine;
+pub use rayon_engine::RayonEngine;
+pub use rayon_limited_engine::RayonLimitedEngine;
 pub use file_discovery::{default_claude_pattern, discover_claude_files, expand_tilde};
